requirements.txt
asyncio
binance-connector
cryptography
discord.py==1.7.3
numpy
pandas
requests
ta==0.10.1
yfinance==0.1.74